admin-panel/requirements.txt
Django==2.2
djangorestframework==3.9.4
requests==2.21.0